+++ CloudShell.ResourceDefinitions/ResourceResolver.cs
@@ -694,6 +694,7 @@ public static class ResourceDefinitionDiagnosticCodes
     public const string ResourceTypeProviderMissing = "resourceDefinition.resourceTypeProviderMissing";
     public const string DuplicateResourceDefinition = "resourceDefinition.duplicateResourceDefinition";
     public const string ResourceDependencyMissing = "resourceDefinition.resourceDependencyMissing";
+    public const string ResourceReferenceTypeMismatch = "resourceDefinition.referenceTypeMismatch";
     public const string ResourceCapabilityReferenceMissing = "resourceDefinition.capabilityReferenceMissing";
     public const string ResourceCapabilityReferenceInvalid = "resourceDefinition.capabilityReferenceInvalid";
     public const string ResourceDefinitionApplyProviderMissing = "resourceDefinition.applyProviderMissing";